Test Cases Added

+++ src/main/java/edu/mum/EmployeeDao.java
@@ -4,22 +4,22 @@ import java.util.List;
 
 import org.ektorp.CouchDbConnector;
 import org.ektorp.CouchDbInstance;
+import org.ektorp.DocumentNotFoundException;
 import org.ektorp.ViewQuery;
 
 public class EmployeeDao {
-	
+
 	private CouchDbInstance dbInstance = DBConnector.getInstance();
 	CouchDbConnector db = null;
-	
+
 	public EmployeeDao() {
 		db = dbInstance.createConnector("employee", true);
 		db.createDatabaseIfNotExists();
 	}
-	
-	public List<Employee> findAll(){
-		ViewQuery view = new ViewQuery().allDocs()
-				.includeDocs(true);
-		return db.queryView(view,Employee.class);
+
+	public List<Employee> findAll() {
+		ViewQuery view = new ViewQuery().allDocs().includeDocs(true);
+		return db.queryView(view, Employee.class);
 	}
 
 	public Employee save(Employee emp) {
@@ -27,22 +27,32 @@ public class EmployeeDao {
 		return emp;
 	}
 
-	public Employee update(Employee emp) {
-		Employee employee = db.get(Employee.class, emp.getId());
+	public boolean update(Employee emp) {
+		Employee employee = findById(emp.getId());
+		if (emp.equals(employee))
+			return false;
 		employee.setName(emp.getName());
 		employee.setSalary(emp.getSalary());
 		employee.setDept(emp.getDept());
 		db.update(employee);
+		return true;
+	}
+
+	public boolean delete(String id) {
+		Employee employee = findById(id);
+		if(employee==null) return false;
+		db.delete(employee);
+		return true;
+	}
+
+	public Employee findById(String id) {
+		Employee emp = null;
+		try {			
+			emp = db.get(Employee.class, id);
+		}catch(DocumentNotFoundException ex) {
+			
+		}
 		return emp;
 	}
 
-	public void delete(String id) {
-		Employee employee = findById(id);
-		db.delete(employee);
-	}
-	
-	public Employee findById(String id) {
-		return db.get(Employee.class, id);
-	}
-	
 }
